New ExtensionPathRector to replace ExtensionPathBase
