+++ config/drupal-9/drupal-9.3-deprecations.php
@@ -2,21 +2,19 @@
 
 declare(strict_types=1);
 
-use DrupalRector\Rector\Deprecation\DrupalGetFilenameRector;
-use DrupalRector\Rector\Deprecation\DrupalGetPathRector;
 use DrupalRector\Rector\Deprecation\FileBuildUriRector;
 use DrupalRector\Rector\Deprecation\FileUrlGenerator;
 use DrupalRector\Rector\Deprecation\FunctionToServiceRector;
 use DrupalRector\Rector\ValueObject\FunctionToServiceConfiguration;
+use DrupalRector\Rector\ValueObject\ExtensionPathConfiguration;
 
 return static function (\Rector\Config\RectorConfig $rectorConfig): void {
     // Change record: https://www.drupal.org/node/2940438.
-    $rectorConfig->ruleWithConfiguration(DrupalGetPathRector::class, [
-            'drupal_rector_notices_as_comments' => '%drupal_rector_notices_as_comments%',
-        ]);
-    $rectorConfig->ruleWithConfiguration(DrupalGetFilenameRector::class, [
-            'drupal_rector_notices_as_comments' => '%drupal_rector_notices_as_comments%',
-        ]);
+    $rectorConfig->ruleWithConfiguration(\DrupalRector\Rector\Deprecation\ExtensionPathRector::class, [
+        'drupal_rector_notices_as_comments' => '%drupal_rector_notices_as_comments%',
+        new ExtensionPathConfiguration('drupal_get_filename', 'getPathname'),
+        new ExtensionPathConfiguration('drupal_get_path', 'getPath'),
+    ]);
 
     // Change record: https://www.drupal.org/node/2940031
     $rectorConfig->rule(FileUrlGenerator\FileCreateUrlRector::class);
