--- src/Rector/Deprecation/Base/ExtensionPathBase.php
@@ -1,75 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace DrupalRector\Rector\Deprecation\Base;
-
-use DrupalRector\Utility\AddCommentTrait;
-use PhpParser\Node;
-use Rector\Core\Contract\Rector\ConfigurableRectorInterface;
-use Rector\Core\Rector\AbstractRector;
-
-abstract class ExtensionPathBase extends AbstractRector implements ConfigurableRectorInterface
-{
-    use AddCommentTrait;
-
-    protected $functionName;
-
-    protected $methodName;
-
-    public function configure(array $configuration): void
-    {
-        $this->configureNoticesAsComments($configuration);
-    }
-
-    public function getNodeTypes(): array
-    {
-        return [
-            Node\Stmt\Expression::class,
-        ];
-    }
-
-    public function refactor(Node $node): ?Node
-    {
-        assert($node instanceof Node\Stmt\Expression);
-        if (!($node->expr instanceof Node\Expr\FuncCall)) {
-            return null;
-        }
-
-        $expr = $node->expr;
-        if ($this->getName($expr->name) !== $this->functionName) {
-            return null;
-        }
-        $args = $expr->getArgs();
-        if (count($args) !== 2) {
-            $this->addDrupalRectorComment($node, "Invalid call to {$this->functionName}, cannot process.");
-            return null;
-        }
-        [$extensionType, $extensionName] = $args;
-
-        $extensionTypeValueType = $this->nodeTypeResolver->getType($extensionType->value);
-        if ($extensionTypeValueType->getConstantStrings()) {
-            $extensionType = $extensionTypeValueType->getValue();
-        }
-
-        if (in_array($extensionType, ['module', 'theme', 'profile', 'theme_engine'], true)) {
-            $serviceName = "extension.list.$extensionType";
-            $methodArgs = [$extensionName];
-        } else {
-            $this->addDrupalRectorComment(
-                $node,
-                'Unsupported extension type encountered, using extension.path.resolver instead of extension.list'
-            );
-            $serviceName = 'extension.path.resolver';
-            $methodArgs = $args;
-        }
-
-        $service = new Node\Expr\StaticCall(
-            new Node\Name\FullyQualified('Drupal'),
-            'service',
-            [new Node\Arg(new Node\Scalar\String_($serviceName))]
-        );
-        $methodName = new Node\Identifier($this->methodName);
-
-        $node->expr = new Node\Expr\MethodCall($service, $methodName, $methodArgs);
-        return $node;
-    }
-}
