--- tests/src/Rector/Deprecation/DrupalGetPathRector/config/configured_rule.php
@@ -1,9 +1,0 @@
-<?php declare(strict_types=1);
-
-use DrupalRector\Rector\Deprecation\DrupalGetPathRector;
-use DrupalRector\Tests\Rector\Deprecation\DeprecationBase;
-use Rector\Config\RectorConfig;
-
-return static function (RectorConfig $rectorConfig): void {
-    DeprecationBase::addClass(DrupalGetPathRector::class, $rectorConfig);
-};
