--- tests/src/Rector/Deprecation/DrupalGetPathRector/fixture/invalid_call.php.bak
@@ -1,15 +1,0 @@
-<?php
-
-function simple_example() {
-    drupal_get_filename('module', 'node', 'baz', 'bar');
-}
-?>
------
-<?php
-
-function simple_example() {
-    // TODO: Drupal Rector Notice: Please delete the following comment after you've made any necessary changes.
-    // Invalid call to drupal_get_filename, cannot process.
-    drupal_get_filename('module', 'node', 'baz', 'bar');
-}
-?>
