+++ tests/src/Rector/Deprecation/ExtensionPathRector/ExtensionPathRectorTest.php
@@ -1,12 +1,12 @@
 <?php declare(strict_types=1);
 
-namespace DrupalRector\Tests\Rector\Deprecation\DrupalGetPathRector;
+namespace DrupalRector\Tests\Rector\Deprecation\ExtensionPathRector;
 
 use Iterator;
 use Rector\Testing\PHPUnit\AbstractRectorTestCase;
 use Symplify\SmartFileSystem\SmartFileInfo;
 
-class DrupalGetPathRectorTest extends AbstractRectorTestCase {
+class ExtensionPathRectorTest extends AbstractRectorTestCase {
 
     /**
      * @covers ::refactor
