+++ tests/src/Rector/Deprecation/ExtensionPathRector/fixture/basic.php.inc
@@ -4,6 +4,9 @@ function simple_example() {
     drupal_get_filename('module', 'node');
     drupal_get_filename('theme', 'seven');
     drupal_get_filename('profile', 'standard');
+    drupal_get_path('module', 'node');
+    drupal_get_path('theme', 'seven');
+    drupal_get_path('profile', 'standard');
 }
 ?>
 -----
@@ -13,5 +16,8 @@ function simple_example() {
     \Drupal::service('extension.list.module')->getPathname('node');
     \Drupal::service('extension.list.theme')->getPathname('seven');
     \Drupal::service('extension.list.profile')->getPathname('standard');
+    \Drupal::service('extension.list.module')->getPath('node');
+    \Drupal::service('extension.list.theme')->getPath('seven');
+    \Drupal::service('extension.list.profile')->getPath('standard');
 }
 ?>
