+++ tests/src/Rector/Deprecation/ExtensionPathRector/fixture/invalid_call.php.bak
