+++ tests/src/Rector/Deprecation/ExtensionPathRector/fixture/unknown_extension_type.php.inc
@@ -2,6 +2,7 @@
 
 function simple_example() {
     drupal_get_filename('library', 'node');
+    drupal_get_path('library', 'node');
 }
 ?>
 -----
@@ -11,5 +12,8 @@ function simple_example() {
     // TODO: Drupal Rector Notice: Please delete the following comment after you've made any necessary changes.
     // Unsupported extension type encountered, using extension.path.resolver instead of extension.list
     \Drupal::service('extension.path.resolver')->getPathname('library', 'node');
+    // TODO: Drupal Rector Notice: Please delete the following comment after you've made any necessary changes.
+    // Unsupported extension type encountered, using extension.path.resolver instead of extension.list
+    \Drupal::service('extension.path.resolver')->getPath('library', 'node');
 }
 ?>
